Implement manual mode x-axis movement

+++ xmc4500/APP/protocol.h
@@ -16,8 +16,8 @@
 #include <stdlib.h>
 
 typedef struct{
-    int8_t x_axis;
-    int8_t y_axis;
+    int16_t x_axis;
+    int16_t y_axis;
     uint8_t z_axis;         //z_axis = 0 -> pen up (no painting), z_axis = 1 -> pen down (painting)
     uint8_t pos_mode;       //pos_mode = 0 -> absolute positioning, pos_mode = 1 -> relative addressing
     uint16_t speed;
